--- a/lib/client-options.js
+++ b/lib/client-options.js
@@ -96,8 +96,7 @@ const errors = require("./errors.js");
  * @property {Number} [maxPrepared] Determines the maximum amount of different prepared queries before evicting items
  * from the internal cache. Reaching a high threshold hints that the queries are not being reused, like when
  * hard-coding parameter values inside the queries.
- * Default: ``500``.
- * [TODO: Add support for this field]
+ * Default: `512`.
  * @property {Object} [policies]
  * [TODO: Add support for this field]
  * @property {LoadBalancingPolicy} [policies.loadBalancing] The load balancing policy instance to be used to determine
@@ -330,7 +329,7 @@ function defaultOptions() {
         authProvider: null,
         requestTracker: null,
         metrics: new metrics.DefaultMetrics(),
-        maxPrepared: 500,
+        maxPrepared: null, // Default is 512, defined on the Rust side
         refreshSchemaDelay: 1000,
         isMetadataSyncEnabled: true,
         prepareOnAllHosts: true,
@@ -668,6 +667,7 @@ function setRustOptions(options) {
     rustOptions.applicationName = options.applicationName;
     rustOptions.applicationVersion = options.applicationVersion;
     rustOptions.keyspace = options.keyspace;
+    rustOptions.cacheSize = options.maxPrepared;
     if (options.credentials) {
         rustOptions.credentialsUsername = options.credentials.username;
         rustOptions.credentialsPassword = options.credentials.password;