@adespawn 2025-04-16
I will try to run benchmarks for this PR in a moment